Announced official hiatus via repository's readme
This mod, as of this moment, is on official hiatus as far as 1.10 is concerned. I may work on 1.12 sometime in the near future, but that's dependent on the finding time and energy with an already fairly busy schedule due to college. If anyone has any questions about this, they may reach me on the Modded Minecraft discord server (http://discord.com/moddedminecraft). My user tag is Twisted_Code#8986. Thanks to any and all that may be looking forward to this's completion mod like I am!